Added Load More feature
Added a "Load more" button in Posts.
In App, I created a loadMorePosts function that I pass down to Posts as a prop, to be called when the "Load more" button is pressed.

In loadMorePosts(), I set the new subset of posts to be shown, based on the newly calculated index.

I used the useEffect hook to watch for a change in allFilteredPosts and reset the subset of posts to be shown.